adds conditional return for event-page, where if both flaked a both flaked gif is shown
